+++ src/FakeXrmEasy.Core/XrmFakedContext.Crud.cs
@@ -109,49 +109,70 @@ namespace FakeXrmEasy
             {
                 throw new InvalidOperationException("The entity must not be null");
             }
-            e = e.Clone(e.GetType());
-            var reference = e.ToEntityReferenceWithKeyAttributes();
-            e.Id = GetRecordUniqueId(reference);
+            var clone = e.Clone(e.GetType());
+            var reference = clone.ToEntityReferenceWithKeyAttributes();
+            clone.Id = GetRecordUniqueId(reference);
 
             // Update specific validations: The entity record must exist in the context
-            if (ContainsEntity(e.LogicalName, e.Id))
+            if (!ContainsEntity(clone.LogicalName, clone.Id))
             {
-                var integrityOptions = GetProperty<IIntegrityOptions>();
-
-                var table = Db.GetTable(e.LogicalName);
-
-                // Add as many attributes to the entity as the ones received (this will keep existing ones)
-                var cachedEntity = table.GetById(e.Id);
-                foreach (var sAttributeName in e.Attributes.Keys.ToList())
-                {
-                    var attribute = e[sAttributeName];
-                    if (attribute == null)
-                    {
-                        cachedEntity.Attributes.Remove(sAttributeName);
-                    }
-                    else if (attribute is DateTime)
-                    {
-                        cachedEntity[sAttributeName] = ConvertToUtc((DateTime)e[sAttributeName]);
-                    }
-                    else
-                    {
-                        if (attribute is EntityReference && integrityOptions.ValidateEntityReferences)
-                        {
-                            var target = (EntityReference)e[sAttributeName];
-                            attribute = ResolveEntityReference(target);
-                        }
-                        cachedEntity[sAttributeName] = attribute;
-                    }
-                }
-
-                // Update ModifiedOn
-                cachedEntity["modifiedon"] = DateTime.UtcNow;
-                cachedEntity["modifiedby"] = CallerProperties.CallerId;
+                throw FakeOrganizationServiceFaultFactory.New($"{clone.LogicalName} with Id {clone.Id} Does Not Exist");
             }
-            else
+            
+            var integrityOptions = GetProperty<IIntegrityOptions>();
+
+            var table = Db.GetTable(e.LogicalName);
+
+            // Add as many attributes to the entity as the ones received (this will keep existing ones)
+            var cachedEntity = table.GetById(clone.Id);
+            foreach (var sAttributeName in clone.Attributes.Keys.ToList())
             {
-                // The entity record was not found, return a CRM-ish update error message
-                throw FakeOrganizationServiceFaultFactory.New($"{e.LogicalName} with Id {e.Id} Does Not Exist");
+                var attribute = clone[sAttributeName];
+                if (attribute == null)
+                {
+                    cachedEntity.Attributes.Remove(sAttributeName);
+                }
+                else if (attribute is DateTime)
+                {
+                    cachedEntity[sAttributeName] = ConvertToUtc((DateTime)clone[sAttributeName]);
+                }
+                else
+                {
+                    if (attribute is EntityReference && integrityOptions.ValidateEntityReferences)
+                    {
+                        var target = (EntityReference) clone[sAttributeName];
+                        attribute = ResolveEntityReference(target);
+                    }
+                    cachedEntity[sAttributeName] = attribute;
+                }
+            }
+
+            // Update ModifiedOn
+            cachedEntity["modifiedon"] = DateTime.UtcNow;
+            cachedEntity["modifiedby"] = CallerProperties.CallerId;
+
+            if (clone.RelatedEntities.Count > 0)
+            {
+                foreach (var relationshipSet in clone.RelatedEntities)
+                {
+                    var relationship = relationshipSet.Key;
+
+                    var entityReferenceCollection = new EntityReferenceCollection();
+
+                    foreach (var relatedEntity in relationshipSet.Value.Entities)
+                    {
+                        UpdateEntity(relatedEntity);
+                        entityReferenceCollection.Add(new EntityReference(relatedEntity.LogicalName, relatedEntity.Id));
+                    }
+
+                    var request = new AssociateRequest
+                    {
+                        Target = clone.ToEntityReference(),
+                        Relationship = relationship,
+                        RelatedEntities = entityReferenceCollection
+                    };
+                    _service.Execute(request);
+                }
             }
         }
 
@@ -392,9 +413,9 @@ namespace FakeXrmEasy
 
             AddEntityWithDefaults(clone, false);
 
-            if (e.RelatedEntities.Count > 0)
+            if (clone.RelatedEntities.Count > 0)
             {
-                foreach (var relationshipSet in e.RelatedEntities)
+                foreach (var relationshipSet in clone.RelatedEntities)
                 {
                     var relationship = relationshipSet.Key;